Move api tests to api_tests subfolder
