+++ api_tester/lib/api_tests_import.dart
@@ -1,16 +1,16 @@
 // ignore_for_file: unused_import
-import 'package:api_tester/errors_api_test.dart';
-import 'package:api_tester/models/customer_info_wrapper_api_test.dart';
-import 'package:api_tester/models/entitlement_info_wrapper_api_test.dart';
-import 'package:api_tester/models/entitlement_infos_wrapper_api_test.dart';
-import 'package:api_tester/models/introductory_price_api_test.dart';
-import 'package:api_tester/models/offering_wrapper_api_test.dart';
-import 'package:api_tester/models/offerings_wrapper_api_test.dart';
-import 'package:api_tester/models/package_wrapper_api_test.dart';
-import 'package:api_tester/models/promotional_offer_api_test.dart';
-import 'package:api_tester/models/purchase_configuration_api_test.dart';
-import 'package:api_tester/models/store_api_test.dart';
-import 'package:api_tester/models/store_product_discount_api_test.dart';
-import 'package:api_tester/models/store_product_wrapper_api_test.dart';
-import 'package:api_tester/models/store_transaction_api_test.dart';
-import 'package:api_tester/purchases_flutter_api_test.dart';
+import 'package:api_tester/api_tests/errors_api_test.dart';
+import 'package:api_tester/api_tests/models/customer_info_wrapper_api_test.dart';
+import 'package:api_tester/api_tests/models/entitlement_info_wrapper_api_test.dart';
+import 'package:api_tester/api_tests/models/entitlement_infos_wrapper_api_test.dart';
+import 'package:api_tester/api_tests/models/introductory_price_api_test.dart';
+import 'package:api_tester/api_tests/models/offering_wrapper_api_test.dart';
+import 'package:api_tester/api_tests/models/offerings_wrapper_api_test.dart';
+import 'package:api_tester/api_tests/models/package_wrapper_api_test.dart';
+import 'package:api_tester/api_tests/models/promotional_offer_api_test.dart';
+import 'package:api_tester/api_tests/models/purchase_configuration_api_test.dart';
+import 'package:api_tester/api_tests/models/store_api_test.dart';
+import 'package:api_tester/api_tests/models/store_product_discount_api_test.dart';
+import 'package:api_tester/api_tests/models/store_product_wrapper_api_test.dart';
+import 'package:api_tester/api_tests/models/store_transaction_api_test.dart';
+import 'package:api_tester/api_tests/purchases_flutter_api_test.dart';
